requirements.txt
agora-token-builder==1.0.0
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appier==1.32.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
aws-xray-sdk==2.14.0
backoff==2.2.1
bcrypt==4.0.1
billiard==4.2.0
boto3==1.34.74
botocore==1.34.74
CacheControl==0.14.0
cachetools==5.3.3
camel-converter==3.1.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
executing==2.1.0
fastapi==0.110.0
firebase-admin==6.5.0
google-api==0.1.12
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
jose==1.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
moto==5.0.17
msgpack==1.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.57.2
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-boto3sqs==0.36b0
opentelemetry-instrumentation-botocore==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-pymongo==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-sqlalchemy==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-starlette==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-urllib3==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-sdk-extension-aws==2.0.1
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
packaging==24.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.0
pyparsing==3.1.2
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.25.3
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
xmltodict==0.14.2
xraysink==1.6.2
zipp==3.20.2
